Remove auto static route gen for originations
This commit makes a change to the way that we build static routes,
specifically originated routes.

If you're originating a route in BIRD, you typically need a static route
definition to get the route into BIRD's RIB. Previous to this pull
request, this was done automatically, but now this must be done
manually.

The reason for this is so that you can override the static route
definitions of originated routes. In some cases you may not want the
route to be `unreachable`, perhaps you want it to be forwarded via
another address recursively. This change makes that possible.

Users will need to ensure that they have static routes configured after
this pull request.